+++ notebook/jungining/3장 템플릿/3.4 컨텍스트와 DI.md
@@ -1,0 +1,82 @@
+# 3.4 컨텍스트와 DI
+## 3.4.1 JdbcContext 분리
+전략 패턴의 구조로 보자면, UserDao는 다음과 같다.
+ - UserDao의 메소드 : 클라이언트
+ - 익명 내부 클래스로 만들어지는 것 : 개별적인 전략
+ - jdbcContextWithStatementStrategy() 메소드 : 컨텍스트 
+ 
+ 컨텍스트 메소드는 UserDao 내의 PreparedStatement를 실행하는 기능을 가진 메소드에서 공유할 수 있다. 그런데 JDBC의 일반적인 작업 흐름을 담고 있는 jdbcContextWithStatementStrategy()는 다른 DAO에서도 사용 가능하다. 독립시켜보자!
+
+
+### 클래스 분리
+분리해서 만들 클래스의 이름은 JdbcContext라고 하자. JdbcContext에 UserDao안에 있는 컨텍스트 메소드를 workWithStatementStrategy()라는 이름으로 옮겨놓는다. DataSource가 필요한 곳이 달라졌다. UserDao가 아니라 JdbcContext안으로 DataSource 타입 빈을 DI받을 수 있도록 준비해놓는다.
+```java
+public class JdbcContext {
+    private DataSource dataSource;
+    public void setDataSource(DataSource dataSource){
+        this.dataSource = dataSource;
+    }
+    public void workWithStatementStrategy(StatementStrategy stmt) throws SQLException{
+       // 기존 jdbcContextWithStatementStrategy()내에 있던 내용이 들어간다.
+    }
+}
+```
+
+이렇게 하면 UserDao가 분리된 JdbcContext를 DI받아 사용할 수 있다.
+
+```java
+public void deleteAll() throws SQLException{
+    jdbcContext.workWithStatementStrategy((new StatementStrategy() {...});
+}
+```
+### 빈 의존관계 변경
+
+**스프링의 DI는 인터페이스를 사이에 두고 의존 클래스를 바꿔서 사용하도록 하는게 목적이다.** 
+스프링의 빈 설정은 클래스 레벨이 아니라 런타임 시에 만들어지는 오브젝트 레벨의 의존관계에 따라 정의된다. 기존에는 userDao빈이 dataSource빈을 직접 의존했지만 이제는 jdbcContext빈이 그 사이에 끼게 된다.
+
+![image](https://user-images.githubusercontent.com/11023497/80294076-a4587880-87a0-11ea-8df5-d4a40f43a4c0.png)
+
+위 의존관계를 따라 xml파일을 수정하자.
+```xml
+<bean id="userDao" class="user.dao.UserDao">
+    <property name="dataSource" ref="dataSource"/>
+    <property name="jdbcContext" ref="jdbcContext"/>
+</bean>
+
+<bean id="jdbcContext" class="user.dao.JdbcContext">
+    <property name="dataSource" ref="dataSource"/>
+</bean>
+```
+이제 JdbcContext를 UserDao로부터 완전히 분리시켜 DI를 통해 연결될 수 있도록 설정을 마쳤다.
+
+## 3.4.2 JdbcConext의 특별한 DI
+### 스프링 빈으로 DI
+
+분리된 클래스인 JdbcContext가 DAO와 인터페이스를 사용하지 않는 직접적인 의존관계를 갖는다. 그러나 DI를 넓게 보면 JdbcContext를 스프링을 이용해 인젝션한다 해도 DI의 기본을 따르고 있다고 볼 수 있다.
+
+인터페이스를 통해 자유롭게 변경이 가능하진 않지만 빈으로 만들어져 스프링 DI를 적용해야 할 이유를 생각해보자.
+
+- 분리된 클래스가 싱글톤 레지스트리에서 관리되는 싱글톤 빈이 되기 때문이다.
+- 분리된 클래스가 DI를 통해 다른 빈에 의존하기 때문이다.
+
+인터페이스가 없다는건 강한 의존 결합된다는 의미다. 다른 방식을 사용해야하는 경우 의존을 갖는 클래스가 통째로 바뀌어야 하므로 굳이 인터페이스를 쓸 필요가 없다. 
+
+### 코드를 이용하는 수동 DI
+
+스프링 빈을 사용하는 방법 대신 클라이언트 클래스에서 직접 DI를 적용하는 방법이 있다. 조금만 타협해서 DAO마다 하나의 JdbcContext 오브젝트를 갖고 있게 하는 것이다. DAO 메서드에서 만들어서 사용하면 수백만의 오브젝트가 생성되겠지만, DAO마다 하나씩이라면 크게 문제가 되진 않는다.
+
+JdbcContext가 DataSource 빈을 의존하고 있기 때문에, 직접 JdbcContext를 사용하더라도 DataSource 빈을 주입받아야한다. 이런 경우에 클라이언트에게 JdbcContext의 제어권을 주고 DI 컨테이너처럼 동작하게 만들면 된다.
+
+설정파일에서 JdbcContext빈을 제거하고, setJdbcContext()도 제거한다. 아래와 같이 setDataSource()를 수정해준다.
+```java
+public void setDataSource(DataSource d){
+    // jdbcContext 생성(IoC)
+    this.jdbcContext = new JdbcContext();
+    
+    // 의존 오브젝트 주입(DI)
+    this.jdbcContext.setDataSource(d);
+    this.dataSource = d;
+}
+```
+
+이 방법은 긴밀한 관계를 갖는 DAO와 JdbcContext를 굳이 어색하게 빈으로 분리하지 않고 내부에서 직접 사용하면서도 다른 오브젝트에 대한 DI를 적용할 수 있다. 어색한 관계가 설정파일에 드러나지 않지만, JdbcContext를 여러 오브젝트가 사용하더라도 싱글톤으로 만들 수 없고, DI 작업을 위한 부가적인 코드가 필요하다는 단점이 있다.
