+++ source/junginig/Springbook/src/user/dao/JdbcContext.java
@@ -1,0 +1,30 @@
+package user.dao;
+
+import javax.sql.DataSource;
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+
+public class JdbcContext {
+    private DataSource dataSource;
+
+    public void setDataSource(DataSource dataSource){
+        this.dataSource = dataSource;
+    }
+
+    public void workWithStatementStrategy(StatementStrategy stmt) throws SQLException{
+        Connection c = null;
+        PreparedStatement ps = null;
+
+        try {
+            c = dataSource.getConnection();
+            ps = stmt.makePreparedStatement(c);
+            ps.executeUpdate();
+        } catch(SQLException e){
+            throw e;
+        } finally {
+            if(ps!=null){ try{  ps.close();  } catch (SQLException e){ } }
+            if(c!=null){ try{  c.close();  } catch (SQLException e){ } }
+        }
+    }
+}
