+++ source/junginig/Springbook/src/user/dao/UserDao.java
@@ -13,13 +13,14 @@ import javax.sql.DataSource;
 
 public class UserDao {
     private DataSource dataSource;
-
-    public void setDataSource(DataSource dataSource){
-        this.dataSource = dataSource;
+    private JdbcContext jdbcContext;
+    public void setDataSource(DataSource d){
+        this.jdbcContext = new JdbcContext();
+        this.jdbcContext.setDataSource(d);
+        this.dataSource = d;
     }
-
     public void add(final User user) throws SQLException {
-        jdbcContextWithStatementStrategy( new StatementStrategy() {
+        jdbcContext.workWithStatementStrategy( new StatementStrategy() {
             @Override
             public PreparedStatement makePreparedStatement(Connection c) throws SQLException {
                 PreparedStatement ps = c.prepareStatement(
@@ -61,12 +62,12 @@ public class UserDao {
 
 
     public void deleteAll() throws SQLException{
-        jdbcContextWithStatementStrategy(new StatementStrategy() {
+        jdbcContext.workWithStatementStrategy((new StatementStrategy() {
             @Override
             public PreparedStatement makePreparedStatement(Connection c) throws SQLException {
                return c.prepareStatement("delete from users");
             }
-        });
+        }));
     }
 
     public int getCount() throws SQLException  {
@@ -98,22 +99,6 @@ public class UserDao {
                     c.close();
                 } catch (SQLException e) { }
             }
-        }
-    }
-
-    public void jdbcContextWithStatementStrategy(StatementStrategy stmt) throws SQLException{
-        Connection c = null;
-        PreparedStatement ps = null;
-
-        try {
-            c = dataSource.getConnection();
-            ps = stmt.makePreparedStatement(c);
-            ps.executeUpdate();
-        } catch(SQLException e){
-            throw e;
-        } finally {
-            if(ps!=null){ try{  ps.close();  } catch (SQLException e){ } }
-            if(c!=null){ try{  c.close();  } catch (SQLException e){ } }
         }
     }
 }
